Upgrade FEMU from QEMU 9.0.1 to 10.1.0 with device registration fixes
  - Update FEMU to use QEMU 10.1.0 as base platform
  - Update device registration to use modern QEMU APIs
  - Maintain compatibility across all SSD emulation modes (BlackBox, WhiteBox, ZNS, NoSSD)

+++ roms/openbios/arch/ppc/ofmem.c
@@ -131,7 +131,7 @@ void ofmem_arch_map_pages(ucell phys, ucell virt, ucell size, ucell mode)
 /************************************************************************/
 
 void *
-malloc( int size )
+malloc( size_t size )
 {
 	return ofmem_malloc(size);
 }
