+++ roms/openbios/include/libc/stdlib.h
@@ -16,7 +16,7 @@
 #ifndef _H_STDLIB
 #define _H_STDLIB
 
-extern void	*malloc( int size );
+extern void	*malloc( size_t size );
 extern void	free( void *ptr );
 extern void	*realloc( void *ptr, size_t size );
